Add `static_panel_nav` option
My project had a requirement to ensure the next/prev arrows don't
disappear. We believed this is better for accessibility. I created a new
option called `static_panel_nav`. When set to TRUE, `static_panel_nav`
will always display the next/prev arrows on the screen. The option is
FALSE by default, to match pre-existing behaviour of the library.